# bls-server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BLS-KOX API",
    "version": "1.0.0",
    "description": "BLS-KOX 多租户后台管理系统 API 文档\n\n## 认证\n在请求头添加 `Authorization: Bearer <token>`\n登录接口获取 token：`POST /api/auth/login`\n\n## 默认账号\n- 账号: `superadmin`  密码: `123456`  租户: `000000`\n- 账号: `admin`       密码: `123456`  租户: `100000`",
    "contact": {
      "name": "BLS-KOX",
      "url": "https://github.com/npcxl/BLS-KOX"
    }
  },
  "servers": [
    {
      "url": "http://localhost:6001",
      "description": "本地开发"
    },
    {
      "url": "/api",
      "description": "Nginx 反代（生产）"
    }
  ],
  "paths": {
...
    "/api/system/config/list": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "系统参数 - 分页列表查询",
        "operationId": "get_api_system_config_list",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "config_id": {
                            "type": "string",
                            "description": "参数ID"
                          },
                          "tenant_id": {
                            "type": "string",
                            "description": "租户ID"
                          },
                          "config_key": {
                            "type": "string",
                            "description": "参数键名"
                          },
                          "config_value": {
                            "type": "string",
                            "description": "参数键值"
                          },
                          "config_name": {
                            "type": "string",
                            "description": "参数名称"
                          },
                          "config_type": {
                            "type": "string",
                            "description": "sys/theme/dict"
                          },
                          "status": {
                            "type": "string",
                            "description": "0正常 1停用"
                          },
                          "remark": {
                            "type": "string",
                            "description": "备注"
                          },
                          "deleted": {
                            "type": "integer",
                            "description": "逻辑删除"
                          },
                          "create_time": {
                            "type": "string",
                            "description": "string"
                          },
                          "update_time": {
                            "type": "string",
                            "description": "string"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 100
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "所需权限: system:config:list",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "页码，默认 1"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "每页条数，默认 10，最大 100"
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "关键词（config_key, config_name）"
          },
          {
            "name": "config_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "精确过滤（白名单字段）"
          },
          {
            "name": "config_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "精确过滤（白名单字段）"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "精确过滤（白名单字段）"
          }
        ]
      }
    },
...
    "/api/system/global-search/config/list": {
      "get": {
        "tags": [
          "global-search"
        ],
        "summary": "global-search - 分页列表",
        "operationId": "get_api_system_global-search_config_list",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/global-search_get__api_system_global-search_config_list"
              }
            }
          }
        }
      }
    },
    "/api/system/global-search/config/save": {
      "post": {
        "tags": [
          "global-search"
        ],
        "summary": "global-search - 保存配置",
        "operationId": "post_api_system_global-search_config_save",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/global-search_post__api_system_global-search_config_save"
              }
            }
          }
        }
      }
    },
...
    "/api/system/page-config/list": {
      "get": {
        "tags": [
          "page-config"
        ],
        "summary": "page-config - 分页列表",
        "operationId": "get_api_system_page-config_list",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/page-config_get__api_system_page-config_list"
              }
            }
          }
        }
      }
    },
    "/api/system/page-config/page/{pageCode}": {
      "get": {
        "tags": [
          "page-config"
        ],
        "summary": "page-config - 查询 :pageCode",
        "operationId": "get_api_system_page-config_page_pageCode_",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "pageCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "pageCode"
          }
        ]
      },
      "delete": {
        "tags": [
          "page-config"
        ],
        "summary": "page-config - 删除 :pageCode",
        "operationId": "delete_api_system_page-config_page_pageCode_",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          },
          {
            "name": "pageCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "pageCode"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/page-config_delete__api_system_page-config_page_pageCode_"
              }
            }
          }
        }
      }
    },
    "/api/system/page-config/page/{pageCode}/columns": {
      "get": {
        "tags": [
          "page-config"
        ],
        "summary": "page-config - 查询 columns",
        "operationId": "get_api_system_page-config_page_pageCode_columns",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "pageCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "pageCode"
          }
        ]
      }
    },
    "/api/system/page-config/save": {
      "post": {
        "tags": [
          "page-config"
        ],
        "summary": "page-config - 保存配置",
        "operationId": "post_api_system_page-config_save",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/page-config_post__api_system_page-config_save"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "global-search_get__api_system_global-search_config_list": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "global-search_post__api_system_global-search_config_save": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          },
          "searchId": {
            "type": "string",
            "description": "searchId"
          },
          "updateTable": {
            "type": "string",
            "description": "updateTable"
          },
          "insertInto": {
            "type": "string",
            "description": "insertInto"
          }
        }
      },
...
      "page-config_get__api_system_page-config_list": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "page-config_delete__api_system_page-config_page_pageCode_": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "page-config_post__api_system_page-config_save": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录获取的 accessToken"
      }
    }
  },
...
}